Work in holidays:
fix managed 2
completed Group Requirement 22
started Group Requirement 23
completed Individual Student Requirement 4: 11 and 12
started Individual Requirement Student 4: 13

+++ src/main/java/acme/entities/offer/Offer.java
@@ -56,8 +56,8 @@ public class Offer extends AbstractEntity {
 
 
 	@Transient
-	public Double getAvailabilityPeriod() {
-		return (double) (MomentHelper.computeDuration(this.availabilityPeriodFinal, this.availabilityPeriodInitial).toMinutes() / 60);
+	public Long getAvailabilityPeriod() {
+		return MomentHelper.computeDuration(this.availabilityPeriodFinal, this.availabilityPeriodInitial).toMinutes() / 60;
 	}
 
 }
